+++ src/main/java/com/lickhunter/spotbot/models/sentiments/Btc.java
@@ -44,7 +44,7 @@ import java.util.Map;
 public class Btc {
 
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
     @JsonProperty("s")
     private String s;
     @JsonProperty("n")
@@ -52,7 +52,7 @@ public class Btc {
     @JsonProperty("p")
     private Double p;
     @JsonProperty("p_btc")
-    private Integer pBtc;
+    private Long pBtc;
     @JsonProperty("v")
     private Double v;
     @JsonProperty("vt")
@@ -66,31 +66,31 @@ public class Btc {
     @JsonProperty("gs")
     private Double gs;
     @JsonProperty("ss")
-    private Integer ss;
+    private Long ss;
     @JsonProperty("as")
     private Double as;
     @JsonProperty("bl")
-    private Integer bl;
+    private Long bl;
     @JsonProperty("br")
-    private Integer br;
+    private Long br;
     @JsonProperty("sp")
-    private Integer sp;
+    private Long sp;
     @JsonProperty("na")
-    private Integer na;
+    private Long na;
     @JsonProperty("md")
-    private Integer md;
+    private Long md;
     @JsonProperty("t")
-    private Integer t;
+    private Long t;
     @JsonProperty("r")
-    private Integer r;
+    private Long r;
     @JsonProperty("yt")
-    private Integer yt;
+    private Long yt;
     @JsonProperty("sv")
-    private Integer sv;
+    private Long sv;
     @JsonProperty("u")
-    private Integer u;
+    private Long u;
     @JsonProperty("c")
-    private Integer c;
+    private Long c;
     @JsonProperty("sd")
     private Double sd;
     @JsonProperty("d")
@@ -98,21 +98,21 @@ public class Btc {
     @JsonProperty("cr")
     private Double cr;
     @JsonProperty("acr")
-    private Integer acr;
+    private Long acr;
     @JsonProperty("tc")
-    private Integer tc;
+    private Long tc;
     @JsonProperty("categories")
     private String categories;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
     @JsonProperty("id")
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
     @JsonProperty("id")
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -147,12 +147,12 @@ public class Btc {
     }
 
     @JsonProperty("p_btc")
-    public Integer getpBtc() {
+    public Long getpBtc() {
         return pBtc;
     }
 
     @JsonProperty("p_btc")
-    public void setpBtc(Integer pBtc) {
+    public void setpBtc(Long pBtc) {
         this.pBtc = pBtc;
     }
 
@@ -217,12 +217,12 @@ public class Btc {
     }
 
     @JsonProperty("ss")
-    public Integer getSs() {
+    public Long getSs() {
         return ss;
     }
 
     @JsonProperty("ss")
-    public void setSs(Integer ss) {
+    public void setSs(Long ss) {
         this.ss = ss;
     }
 
@@ -237,112 +237,112 @@ public class Btc {
     }
 
     @JsonProperty("bl")
-    public Integer getBl() {
+    public Long getBl() {
         return bl;
     }
 
     @JsonProperty("bl")
-    public void setBl(Integer bl) {
+    public void setBl(Long bl) {
         this.bl = bl;
     }
 
     @JsonProperty("br")
-    public Integer getBr() {
+    public Long getBr() {
         return br;
     }
 
     @JsonProperty("br")
-    public void setBr(Integer br) {
+    public void setBr(Long br) {
         this.br = br;
     }
 
     @JsonProperty("sp")
-    public Integer getSp() {
+    public Long getSp() {
         return sp;
     }
 
     @JsonProperty("sp")
-    public void setSp(Integer sp) {
+    public void setSp(Long sp) {
         this.sp = sp;
     }
 
     @JsonProperty("na")
-    public Integer getNa() {
+    public Long getNa() {
         return na;
     }
 
     @JsonProperty("na")
-    public void setNa(Integer na) {
+    public void setNa(Long na) {
         this.na = na;
     }
 
     @JsonProperty("md")
-    public Integer getMd() {
+    public Long getMd() {
         return md;
     }
 
     @JsonProperty("md")
-    public void setMd(Integer md) {
+    public void setMd(Long md) {
         this.md = md;
     }
 
     @JsonProperty("t")
-    public Integer getT() {
+    public Long getT() {
         return t;
     }
 
     @JsonProperty("t")
-    public void setT(Integer t) {
+    public void setT(Long t) {
         this.t = t;
     }
 
     @JsonProperty("r")
-    public Integer getR() {
+    public Long getR() {
         return r;
     }
 
     @JsonProperty("r")
-    public void setR(Integer r) {
+    public void setR(Long r) {
         this.r = r;
     }
 
     @JsonProperty("yt")
-    public Integer getYt() {
+    public Long getYt() {
         return yt;
     }
 
     @JsonProperty("yt")
-    public void setYt(Integer yt) {
+    public void setYt(Long yt) {
         this.yt = yt;
     }
 
     @JsonProperty("sv")
-    public Integer getSv() {
+    public Long getSv() {
         return sv;
     }
 
     @JsonProperty("sv")
-    public void setSv(Integer sv) {
+    public void setSv(Long sv) {
         this.sv = sv;
     }
 
     @JsonProperty("u")
-    public Integer getU() {
+    public Long getU() {
         return u;
     }
 
     @JsonProperty("u")
-    public void setU(Integer u) {
+    public void setU(Long u) {
         this.u = u;
     }
 
     @JsonProperty("c")
-    public Integer getC() {
+    public Long getC() {
         return c;
     }
 
     @JsonProperty("c")
-    public void setC(Integer c) {
+    public void setC(Long c) {
         this.c = c;
     }
 
@@ -377,22 +377,22 @@ public class Btc {
     }
 
     @JsonProperty("acr")
-    public Integer getAcr() {
+    public Long getAcr() {
         return acr;
     }
 
     @JsonProperty("acr")
-    public void setAcr(Integer acr) {
+    public void setAcr(Long acr) {
         this.acr = acr;
     }
 
     @JsonProperty("tc")
-    public Integer getTc() {
+    public Long getTc() {
         return tc;
     }
 
     @JsonProperty("tc")
-    public void setTc(Integer tc) {
+    public void setTc(Long tc) {
         this.tc = tc;
     }
